+++ NEWS.md
@@ -4,4 +4,7 @@
 * updated code formating to adhere to the bestpractices() package
 * updated Readme file
 * added smaller test file for the example of DEG_function to test within the allowed timeslot
+* changed print statements to message(), warning() or stop() statements
+* removed 'outputdir' from Maelstrom_Motif2TF, since it was not needed since the data is returned
+  to the S4 objects as assays (and no longer as output files)